# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Demo Locker API",
    "version": "1.0.0",
    "description": "API for Demo Locker, a private-by-default demo/playlist sharing service. Most routes require a Bearer session token; media elements (<audio>/<img>) that cannot send an Authorization header instead accept the same token via a `?token=` query parameter (session token OR a per-playlist share/invite token — both are treated identically). Gated legacy routes never distinguish 'exists but private' from 'does not exist': any access failure returns a generic `{\"error\":\"not found\"}` 404 so private playlist/track/comment IDs cannot be enumerated by probing.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment (Worker or self-hosted)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token returned by POST /auth/signup or POST /auth/login, sent as `Authorization: Bearer <token>`."
      },
      "tokenQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Session token OR a share/invite token, sent as a `?token=` query parameter. Used by media elements (<audio src>, <img src>) that cannot set an Authorization header. Accepted anywhere bearerAuth is accepted on the legacy (non-/public/v1) gated routes."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Playlist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdByMe": {
            "type": "boolean",
            "description": "Attribution, NOT permission: whether the requesting session created this playlist. Computed per request. The raw creator id (playlists.created_by) is never serialized to any client on any route — a locker may hold several collaborators, and none of them may learn another's user id; the same rule as tracks.uploaded_by. Do not build a delete control from this field alone: DELETE /playlists/{id} lets the locker owner delete any playlist, so the owner gets false on every collaborator-created playlist they are nonetheless allowed to delete. The client-side rule is createdByMe OR locker ownership. The server never reads this field back — authorisation is decided from stored columns on every request. False is ambiguous and must not be rendered as \"someone else created this\": it also means no creator is recorded (rows predating the column), or that the requester has no identity to compare, which includes an anonymous share/invite holder."
          },
          "createdByName": {
            "type": [
              "string",
              "null"
            ],
            "description": "Attribution by NAME: who created this playlist. Resolved server-side from the creator's users.display_name — the label the locker owner typed when minting that person's invite — falling back to their email address when they have none (the locker owner has no invite, and so has none). The raw creator id is still never serialized: this is a name, not the id under another key. createdByMe alone could not answer this — it says \"mine / not mine\", which in a locker holding two collaborators never identifies WHICH of them. Null when no creator is recorded (rows predating the column, or a creator since removed), and null for every reader with no locker session: an anonymous share or invite holder is outside the locker and is served no names at all. Present on the caller's own rows too — rendering those as \"you\" is the client's decision. This is ATTRIBUTION, NOT PERMISSION: it says whose work the row is, and grants and gates nothing. Every control on the row behaves exactly as it did before this field existed, and the server never reads it back — authorisation is decided from stored columns on every request. Null means there is nothing to attribute and MUST render as nothing at all, never as \"unknown\" or \"null\"."
          },
          "name": {
            "type": "string"
          },
          "artworkKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "isPublic": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ownerId",
          "createdByMe",
          "name",
          "isPublic",
          "createdAt",
          "updatedAt",
          "createdByName"
        ],
        "description": "A playlist as returned by every route that serves one. created_by is never serialized: it is replaced by the per-request createdByMe boolean, so the shape is identical for a locker session and for an anonymous share/invite holder (who simply always gets false). One serializer, not two — see lib/public-playlist.ts."
      },
      "Track": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "playlistId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          },
          "waveformData": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration": {
            "type": [
              "number",
              "null"
            ]
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "hasStream": {
            "type": "boolean",
            "description": "Whether a streaming rendition exists yet. Replaces originalKey/streamKey, which were bucket coordinates and are no longer exposed to clients."
          },
          "uploadedByMe": {
            "type": "boolean",
            "description": "Attribution, NOT permission: whether the requesting session uploaded this track. Computed per request. The raw uploader id (tracks.uploaded_by) is never serialized to any client on any route — a locker may hold several collaborators, and none of them may learn another's user id. Do not build a delete control from this field alone: DELETE /tracks/{id} lets the locker owner delete any track, so the owner gets false on every collaborator upload they are nonetheless allowed to delete. The client-side rule is uploadedByMe OR locker ownership. (For a collaborator the two now coincide — they may delete exactly the tracks they uploaded — but the field is still attribution that happens to match, not the authority itself.) The server never reads this field back — authorisation is decided from stored columns on every request. False is ambiguous and must not be rendered as \"someone else uploaded this\": it also means no uploader is recorded (rows predating the column, or an uploader since deleted — the column is ON DELETE SET NULL), or that the requester has no identity to compare, which includes an anonymous share/invite holder."
          },
          "uploadedByName": {
            "type": [
              "string",
              "null"
            ],
            "description": "Attribution by NAME: who uploaded this track. The exact counterpart of Playlist.createdByName, with the same resolution (users.display_name, copied from the uploader's invite label at redemption, falling back to their email) and the same rules. The raw uploader id is still never serialized on any route. Null when no uploader is recorded — rows predating the column, an uploader since deleted (ON DELETE SET NULL), and an anonymous edit-share upload, which has no user id to record — and null for every reader with no locker session, who is served no names at all. This is ATTRIBUTION, NOT PERMISSION: it says whose work the row is, and grants and gates nothing. Every control on the row behaves exactly as it did before this field existed, and the server never reads it back — authorisation is decided from stored columns on every request. Null means there is nothing to attribute and MUST render as nothing at all, never as \"unknown\" or \"null\"."
          }
        },
        "required": [
          "id",
          "playlistId",
          "title",
          "position",
          "uploadedAt",
          "hasStream",
          "uploadedByMe",
          "uploadedByName"
        ]
      },
...
      "Share": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "playlistId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string"
          },
          "permission": {
            "type": "string",
            "enum": [
              "listen",
              "edit"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "mintedByMe": {
            "type": "boolean",
            "description": "Attribution, NOT permission: whether the requesting session minted this share link. Computed per request. The raw minter id (shares.created_by) is never serialized to any client on any route — a locker may hold several collaborators, and none of them may learn another's user id; the same rule as tracks.uploaded_by and playlists.created_by. Do not build a control gate from this field: share links are locker-level state, so any member of the locker may revoke or re-permission any link regardless of who minted it. It exists so the owner can tell their own links from a collaborator's — which matters because removing a collaborator deletes every link they minted (shares.created_by is ON DELETE CASCADE). The server never reads this field back — authorisation is decided from stored columns on every request. False is ambiguous: it also means no minter is recorded (a state migration 0004 backfilled away), or that the requester has no identity to compare."
          }
        },
        "required": [
          "id",
          "playlistId",
          "token",
          "permission",
          "createdAt",
          "mintedByMe"
        ]
      },
...
    }
  },
  "paths": {
...
    "/shares": {
      "post": {
        "summary": "Create a share/invite for a playlist",
        "description": "Locker-scoped: any session acting in the playlist's locker — its owner or a collaborator — may call this. Issues a random unguessable token granting `listen` or `edit` access to the playlist, optionally tied to an email for reference.",
        "operationId": "createShare",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "playlistId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "permission": {
                    "type": "string",
                    "enum": [
                      "listen",
                      "edit"
                    ]
                  },
                  "email": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "email"
                  }
                },
                "required": [
                  "playlistId",
                  "permission"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Share created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "share": {
                      "$ref": "#/components/schemas/Share"
                    }
                  },
                  "required": [
                    "share"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing playlistId/permission, or permission not 'listen'/'edit'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Per-playlist share-link limit reached (MAX_SHARE_LINKS). Not MAX_COLLABORATORS — that caps collaborator seats, and the two were split in 0.2.13.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Playlist not in the caller's locker, or does not exist (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/shares/playlist/{playlistId}": {
      "get": {
        "summary": "List shares for a playlist",
        "description": "Locker-scoped: any session acting in the playlist's locker — its owner or a collaborator — may call this.",
        "operationId": "listPlaylistShares",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shares for the playlist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "shares": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Share"
                      }
                    }
                  },
                  "required": [
                    "shares"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Playlist not in the caller's locker, or does not exist (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/shares/{id}": {
      "delete": {
        "summary": "Revoke a share",
        "description": "Locker-scoped: any session acting in the share's playlist's locker — its owner or a collaborator — may revoke it, including a link a different collaborator or the owner created.",
        "operationId": "deleteShare",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired session token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Share does not exist, or its playlist is not in the caller's locker (non-enumerable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/shares/invite/{token}": {
      "get": {
        "summary": "Resolve an invite/share token",
        "description": "Unauthenticated. Unlike the other gated routes, invalid/expired tokens return distinct error messages (this endpoint is the entry point for invite links, not a resource lookup that must resist enumeration).",
        "operationId": "resolveInvite",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist and tracks the invite grants access to.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "permission": {
                      "type": "string",
                      "enum": [
                        "listen",
                        "edit"
                      ]
                    },
                    "playlist": {
                      "$ref": "#/components/schemas/Playlist"
                    },
                    "tracks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Track"
                      }
                    },
                    "accent": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "#fc0",
                        "#f80",
                        "#4af",
                        "#3f6",
                        "#f6a",
                        "#a6f",
                        "#6cf",
                        null
                      ],
                      "description": "The playlist owner's accent colour, so a shared locker keeps its colours for listeners. Null when the owner never set one."
                    }
                  },
                  "required": [
                    "permission",
                    "playlist",
                    "tracks",
                    "accent"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Token does not exist, or the playlist it referenced was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Token exists but is past its expiresAt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  }
}